Refactor Scheduler model
