+++ src/main/kotlin/com/nylas/models/BookingType.kt
@@ -1,0 +1,14 @@
+package com.nylas.models
+
+import com.squareup.moshi.Json
+
+/**
+ * Enum for booking types.
+ */
+enum class BookingType {
+  @Json(name = "booking")
+  BOOKING,
+
+  @Json(name = "organizer-confirmation")
+  ORGANIZER_CONFIRMATION,
+}
