+++ src/main/kotlin/com/nylas/models/ConfirmBookingStatus.kt
@@ -1,0 +1,14 @@
+package com.nylas.models
+
+import com.squareup.moshi.Json
+
+/**
+ * Enum for confirm booking statuses.
+ */
+enum class ConfirmBookingStatus {
+  @Json(name = "confirm")
+  CONFIRM,
+
+  @Json(name = "cancel")
+  CANCEL,
+}
